Move the coins to spend logic out of off-chain DB adapter
